Fix lowercase bug
Fixed a bug where all the string literals were being converted to lowercase along with the code

+++ vm/src/vm.cpp
@@ -121,7 +121,7 @@ void VM::run() {
                 break;
             }
             default:
-                std::cout << "Unknown opcode: " << opcode << std::endl;
+                std::cout << "Unknown opcode: " << (int) opcode << std::endl;
         }
     }
 }
